+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 * Replaces TravisCI with Github Actions
 * Updates for `sxm-player=0.2.1` client
 * Replaces old school Discord commands with proper slash commands
+* Adds pydantic to `sxm_discord`
 
 0.1.0 (2018-12-25)
 ------------------